requirements.txt
appnope==0.1.2
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
Fiona==1.8.20
geopandas==0.9.0
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
munch==2.5.0
numpy==1.21.0
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyarrow==4.0.1
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
pyproj==3.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Shapely==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
